refactor(utils/currency): move currency to utils
