+++ app/javascript/components/item-list-card.vue
@@ -2,7 +2,7 @@
 import { ref } from 'vue';
 import { useNotification } from '@kyvg/vue3-notification';
 import purchasesApi from '../api/purchases';
-import currency from '../filters/currency.js';
+import currency from '../utils/currency.js';
 
 type Props = {
   itemId: number
