+++ app/javascript/utils/currency.js
